+++ flint/src/Flint/Shell/Repl.hs
@@ -14,6 +14,7 @@ import Flint.Shell.Commands.Paths (sampleCommand, showCommand, pshowCommand, fre
 import Flint.Shell.Commands.Solver (solveCommand)
 import Flint.Shell.Commands.WMI (wmisCommand, checkWMICommand)
 import Flint.Shell.Commands.TypeCheck (typecheckCommand)
+import Flint.Shell.Commands.Strings (stringsCommand, stringXrefsCommand)
 import Flint.Shell.Commands.Xrefs (functionsCallingCommand)
 
 import qualified Data.Text as Text
@@ -37,6 +38,8 @@ allCommands =
   , freeUntaggedCommand
   , expandCommand
   , functionsCallingCommand
+  , stringsCommand
+  , stringXrefsCommand
   ]
 
 runShell :: ShellState -> IO ()